+++ src/ts/components/core/combobox/Select.tsx
@@ -8,6 +8,7 @@ import { __BaseInputProps } from "props/input";
 import { ScrollAreaProps } from "props/scrollarea";
 import { StylesApiProps } from "props/styles";
 import React, { useState } from "react";
+import { isInOption } from "../../../utils/combobox";
 
 interface Props
     extends BoxProps,
@@ -50,6 +51,9 @@ const Select = (props: Props) => {
 
     useDidUpdate(() => {
         setOptions(data);
+        if (!isInOption(data, selected)) {
+            setSelected(null);
+        }
     }, [data]);
 
     useDidUpdate(() => {