+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   Time parsing in DateTimePicker. #273
 -   Show MenuItem as disabled when argument is set to True #202
+-   Update selection when options are changed in combobox components - Select, MultiSelect and TagsInput #201
 
 ### Added